+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - `flipperzero::furi::time::Duration` has been renamed to `FuriDuration`
 - `flipperzero::furi::time::Instant` has been renamed to `FuriInstant`
 - `flipperzero_sys::furi::UnsafeRecord` now implements `Clone` to support reference-count use
+- `flipperzero::notification::NotificationService` has been renamed `NotificationApp`
+  to match the underlying C API
 
 ### Removed